fix import weirdness
