+++ client/tsconfig.json
@@ -31,7 +31,7 @@
             "@coursefull": ["./@coursefull/index.d"],
             "@lib/*": ["./lib/*"],
             "@playwright/*": ["./playwright/*"],
-            "@services*": ["./services/*"],
+            "@services/*": ["./services/*"],
             "@services": ["./services/index.ts"],
             "@vitest.setup": ["./vitest.setup.ts"]
         }
